+++ tests/Support/PluginDirectoryGenerator.php
@@ -23,6 +23,10 @@ class PluginDirectoryGenerator
         }
     }
 
+    public static function use(PluginTemplate $pluginTemplate): bool {
+        return File::exists($pluginTemplate->path());
+    }
+
     /**
      * Creates the basic directory structure for a plugin.
      *
@@ -30,36 +34,6 @@ class PluginDirectoryGenerator
      * @return array The directory structure as an associative array, where keys are file/directory
      *               names and values are either file content (string) or nested arrays for directories.
      */
-    public static function defineStructure(Plugin $plugin, array $additionalStructure = [], string $additionalInfoContent = ''): array
-    {
-        $structure = [
-            "package.json" => self::generatePackageJSON($plugin),
-            "CHANGELOG.md" => "# Changelog",
-            "js"           => [
-                'script.js',
-            ],
-            "App"          => [
-                "info.xml" => self::generateInfoXml($plugin, $additionalInfoContent),
-            ],
-        ];
-
-        $existingDirectories = ['js', 'App'];
-        foreach($existingDirectories as $dir) {
-            if(isset($additionalStructure[$dir])) {
-                // If the content is an array we need to merge it with the existing structure.
-                if(is_array($additionalStructure[$dir])) {
-                    $structure[$dir] = array_merge($structure[$dir], $additionalStructure[$dir]);
-                } else {
-                    // If it's not an array, it's a file. We can just add it to the directory.
-                    $structure[$dir][] = $additionalStructure[$dir];
-                }
-                // We unset the additional structure for this directory to avoid processing it again later.
-                unset($additionalStructure[$dir]);
-            }
-        }
-
-        return array_merge($structure, $additionalStructure);
-    }
 
     /**
      * Returns the full path to the plugin directory for a given plugin name.
@@ -77,12 +51,12 @@ class PluginDirectoryGenerator
      * @param {array} $additionalStructure An associative array representing additional files or directories to be added to the plugin structure.
      * @param {string} $additionalInfoContent Additional XML content to be included in the info.xml file of the plugin.
      */
-    public static function mockPluginDirectory(Plugin $plugin, array $additionalStructure = [], string $additionalInfoContent = ''): string
+    public static function mockPluginDirectory(PluginTemplate $template): string
     {
         // Work on a clone so we don't mutate the caller's Plugin instance (array access
         // inside this helper may set attributes on the Eloquent model which causes
         // unexpected side effects such as trying to insert arrays into the DB).
-        $pluginCopy = clone $plugin;
+        $pluginCopy = clone $template->plugin;
 
         $required_fields = ['name', 'uuid', 'version'];
 
@@ -103,9 +77,8 @@ class PluginDirectoryGenerator
             mkdir($pluginDir, 0755, true);
         }
 
-        $structure = self::defineStructure($pluginCopy, $additionalStructure, $additionalInfoContent);
         // Process structure array
-        foreach($structure as $name => $content) {
+        foreach($template->getStructure() as $name => $content) {
             $path = $pluginDir . '/' . $name;
 
             if(is_array($content)) {
@@ -148,36 +121,9 @@ class PluginDirectoryGenerator
 JSON;
     }
 
-    /**
-     * Generates the content for the info.xml file of the plugin based on the plugin's information and any additional XML content provided.
-     *
-     * @return string The generated XML content for the plugin's info.xml file.
-     */
-    public static function generateInfoXml(Plugin $plugin, $additionalText = ""): string
-    {
-        $description = isset($plugin['description']) ? $plugin['description'] : "";
-        $licence     = isset($plugin['licence']) ? $plugin['licence'] : "";
-        $authors     = "";
-        if(! isset($plugin["authors"])) {
-            $plugin["authors"] = [];
-        }
 
-        foreach($plugin["authors"] as $author) {
-            $authors .= "        <author>{$author}</author>\n";
-        }
 
-        return <<<XML
-<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<info>
-    <!-- Must match id in JS SpPS.register call in kebab-case -->
-    <name>{$plugin['name']}</name>
-    <title>{$plugin['name']} Plugin</title>
-    <description>{$description}</description>
-    <version>{$plugin['version']}</version>
-    <licence>{$licence}</licence>
-    <authors>{$authors}</authors>
-    {$additionalText}
-</info>
-XML;
-    }
+
 }
+
+